requirements.txt
alembic==1.7.4
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.2
Flask==2.0.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
numpy==1.21.4
pandas==1.3.4
psycopg2-binary==2.9.1
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
urllib3==1.26.7
Werkzeug==2.0.2
APScheduler==3.0.0
